Add Main Menu
These commits make a main menu for the game. The menu consists of 3 buttons (new game, continue, and settings), each leading to a very basic scene. From these scenes, you can click a "back" button to go back to the main menu. In addition, from the "settings" scene, you can adjust the volume of the music that's playing.